+++ Warehouse/Models/BookResponse.cs
@@ -9,7 +9,7 @@ namespace Warehouse.Models
     {
         public string name { get; set; }
         public int quantity { get; set; }
-        public int price { get; set; }
-        public int discount { get; set; }
+        public double price { get; set; }
+        public double discount { get; set; }
     }
 }
